+++ lib/cu_cp/task_schedulers/cu_up_task_scheduler.h
@@ -25,7 +25,7 @@ class cu_up_task_scheduler
 public:
   explicit cu_up_task_scheduler(timer_manager&        timers_,
                                 task_executor&        exec_,
-                                uint16_t              max_nof_cu_ups,
+                                unsigned              max_nof_cu_ups,
                                 srslog::basic_logger& logger_);
   ~cu_up_task_scheduler() = default;
